# rcn-devops-kit/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TerraFusionBuild RCN Valuation Engine API",
    "description": "API for calculating Replacement Cost New (RCN) values for property assessment and valuation purposes.",
    "version": "1.0.0",
    "contact": {
      "name": "TerraFusionBuild Support",
      "email": "support@terrafusionbuild.com",
      "url": "https://terrafusionbuild.com/support"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://terrafusionbuild.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API Information",
        "description": "Returns information about the API including version and available endpoints.",
        "operationId": "getApiInfo",
        "responses": {
          "200": {
            "description": "API information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiInfo"
                }
              }
            }
          }
        }
      }
    },
    "/rcn/calculate": {
      "post": {
        "summary": "Calculate RCN Value",
        "description": "Calculates the Replacement Cost New (RCN) value for a building based on its characteristics.",
        "operationId": "calculateRcn",
        "requestBody": {
          "description": "Building information for RCN calculation",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RCNInput"
              },
              "examples": {
                "residential": {
                  "summary": "Residential Wood Frame Example",
                  "value": {
                    "use_type": "Residential",
                    "construction_type": "Wood Frame",
                    "sqft": 1800,
                    "year_built": 2010,
                    "quality_class": "B",
                    "locality_index": 1.05,
                    "condition": "Good"
                  }
                },
                "commercial": {
                  "summary": "Commercial Masonry Example",
                  "value": {
                    "use_type": "Commercial",
                    "construction_type": "Masonry",
                    "sqft": 5000,
                    "year_built": 2000,
                    "quality_class": "A",
                    "locality_index": 1.15,
                    "condition": "Average"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RCN calculation results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RCNOutput"
                },
                "example": {
                  "base_cost": 198000.00,
                  "adjusted_rcn": 207900.00,
                  "depreciated_rcn": 166320.00,
                  "depreciation_pct": 20.00
                }
              }
            }
          },
          "400": {
            "description": "Invalid input parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": [
                    {
                      "loc": ["body", "sqft"],
                      "msg": "value is not a valid float",
                      "type": "type_error.float"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "detail": "An error occurred during calculation"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "API name",
            "example": "TerraFusionBuild RCN Valuation Engine"
          },
          "version": {
            "type": "string",
            "description": "API version",
            "example": "1.0.0"
          },
          "description": {
            "type": "string",
            "description": "Brief description of the API",
            "example": "API for calculating Replacement Cost New (RCN) values for property valuation"
          },
          "endpoints": {
            "type": "array",
            "description": "List of available API endpoints",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "description": "Endpoint path",
                  "example": "/rcn/calculate"
                },
                "method": {
                  "type": "string",
                  "description": "HTTP method",
                  "example": "POST"
                },
                "description": {
                  "type": "string",
                  "description": "Endpoint description",
                  "example": "Calculate RCN value for a building"
                }
              }
            }
          }
        }
      },
      "RCNInput": {
        "type": "object",
        "required": [
          "use_type",
          "construction_type",
          "sqft",
          "year_built"
        ],
        "properties": {
          "use_type": {
            "type": "string",
            "description": "Building use type",
            "example": "Residential",
            "enum": ["Residential", "Commercial", "Industrial", "Agricultural"]
          },
          "construction_type": {
            "type": "string",
            "description": "Primary construction material/type",
            "example": "Wood Frame",
            "enum": ["Wood Frame", "Masonry", "Steel Frame", "Concrete"]
          },
          "sqft": {
            "type": "number",
            "format": "float",
            "description": "Total square footage of the building",
            "example": 1800,
            "minimum": 1
          },
          "year_built": {
            "type": "integer",
            "description": "Year the building was constructed",
            "example": 2010,
            "minimum": 1800,
            "maximum": 2025
          },
          "quality_class": {
            "type": "string",
            "description": "Quality class rating of the building",
            "example": "B",
            "enum": ["A+", "A", "B+", "B", "C+", "C", "D+", "D", "E"],
            "default": "B"
          },
          "locality_index": {
            "type": "number",
            "format": "float",
            "description": "Local cost multiplier for the area (1.0 = average)",
            "example": 1.05,
            "minimum": 0.5,
            "maximum": 2.0,
            "default": 1.0
          },
          "condition": {
            "type": "string",
            "description": "Current physical condition of the building",
            "example": "Good",
            "enum": ["Excellent", "Good", "Average", "Fair", "Poor"],
            "default": "Average"
          }
        }
      },
      "RCNOutput": {
        "type": "object",
        "properties": {
          "base_cost": {
            "type": "number",
            "format": "float",
            "description": "Base cost before adjustments ($/sqft × sqft)",
            "example": 198000.00
          },
          "adjusted_rcn": {
            "type": "number",
            "format": "float",
            "description": "Adjusted Replacement Cost New value after locality and quality adjustments",
            "example": 207900.00
          },
          "depreciated_rcn": {
            "type": "number",
            "format": "float",
            "description": "Depreciated RCN value after applying physical, functional, and economic depreciation",
            "example": 166320.00
          },
          "depreciation_pct": {
            "type": "number",
            "format": "float",
            "description": "Total depreciation percentage applied",
            "example": 20.00
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "oneOf": [
              {
                "type": "string",
                "description": "Error message"
              },
              {
                "type": "array",
                "description": "List of validation errors",
                "items": {
                  "type": "object",
                  "properties": {
                    "loc": {
                      "type": "array",
                      "description": "Location of the error",
                      "items": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "integer"
                          }
                        ]
                      }
                    },
                    "msg": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "type": {
                      "type": "string",
                      "description": "Error type"
                    }
                  }
                }
              }
            ]
          }
        }
      }
    }
  }
}